Isolate tests for exception interceptor
