+++ tests/client_test.py
@@ -21,22 +21,15 @@ import logging
 from unittest import TestCase
 from importlib import import_module
 
-import grpc
 from pyfakefs.fake_filesystem_unittest import TestCase as FileTestCase
 
 from google.ads.google_ads import client as Client
-from google.ads.google_ads.errors import GoogleAdsException
 
 latest_version = Client._DEFAULT_VERSION
 valid_versions = Client._VALID_API_VERSIONS
 
-errors_path = 'google.ads.google_ads.{}.proto.errors'.format(latest_version)
 services_path = 'google.ads.google_ads.{}.proto.services'.format(latest_version)
-
-errors = import_module(errors_path)
 services = import_module(services_path)
-
-error_protos = errors.errors_pb2
 google_ads_service_pb2 = services.google_ads_service_pb2
 
 class ModuleLevelTest(TestCase):
@@ -875,153 +868,3 @@ class LoggingInterceptorTest(TestCase):
             interceptor = self._create_test_interceptor()
             result = interceptor._get_fault_message(mock_exception)
             self.assertEqual(result, self._MOCK_TRANSPORT_ERROR_MESSAGE)
-
-
-class ExceptionInterceptorTest(TestCase):
-    """Tests for the google.ads.googleads.client.ExceptionInterceptor class."""
-
-    _MOCK_FAILURE_VALUE = b"\n \n\x02\x08\x10\x12\x1aInvalid customer ID '123'."
-
-    def _create_test_interceptor(self):
-        """Creates and returns an ExceptionInterceptor instance
-
-        Returns:
-            An ExceptionInterceptor instance.
-        """
-        return Client.ExceptionInterceptor()
-
-    def test_init_(self):
-        """Tests that the interceptor initializes properly"""
-        interceptor = self._create_test_interceptor()
-        self.assertEqual(interceptor._RETRY_STATUS_CODES,
-                         (grpc.StatusCode.INTERNAL,
-                          grpc.StatusCode.RESOURCE_EXHAUSTED))
-
-    def test_get_google_ads_failure(self):
-        """Obtains the content of a google ads failure from metadata."""
-        interceptor = self._create_test_interceptor()
-        mock_metadata = ((interceptor._failure_key, self._MOCK_FAILURE_VALUE),)
-        result = interceptor._get_google_ads_failure(mock_metadata)
-        self.assertIsInstance(result, error_protos.GoogleAdsFailure)
-
-    def test_get_google_ads_failure_decode_error(self):
-        """Returns none if the google ads failure cannot be decoded."""
-        interceptor = self._create_test_interceptor()
-        mock_failure_value = self._MOCK_FAILURE_VALUE + b'1234'
-        mock_metadata = ((interceptor._failure_key, mock_failure_value),)
-        result = interceptor._get_google_ads_failure(mock_metadata)
-        self.assertEqual(result, None)
-
-    def test_get_google_ads_failure_no_failure_key(self):
-        """Returns None if an error cannot be found in metadata."""
-        mock_metadata = (('another-key', 'another-val'),)
-        interceptor = self._create_test_interceptor()
-        result = interceptor._get_google_ads_failure(mock_metadata)
-        self.assertEqual(result, None)
-
-    def test_get_google_ads_failure_with_None(self):
-        """Returns None if None is passed."""
-        interceptor = self._create_test_interceptor()
-        result = interceptor._get_google_ads_failure(None)
-        self.assertEqual(result, None)
-
-    def test_handle_grpc_failure(self):
-        """Raises non-retryable GoogleAdsFailures as GoogleAdsExceptions."""
-        mock_error_message = self._MOCK_FAILURE_VALUE
-
-        class MockRpcErrorResponse(grpc.RpcError):
-            def code(self):
-                return grpc.StatusCode.INVALID_ARGUMENT
-
-            def trailing_metadata(self):
-                return ((interceptor._failure_key, mock_error_message),)
-
-            def exception(self):
-                return self
-
-        interceptor = self._create_test_interceptor()
-
-        self.assertRaises(GoogleAdsException,
-                          interceptor._handle_grpc_failure,
-                          MockRpcErrorResponse())
-
-    def test_handle_grpc_failure_retryable(self):
-        """Raises retryable exceptions as-is."""
-        class MockRpcErrorResponse(grpc.RpcError):
-            def code(self):
-                return grpc.StatusCode.INTERNAL
-
-            def exception(self):
-                return self
-
-        interceptor = self._create_test_interceptor()
-
-        self.assertRaises(MockRpcErrorResponse,
-                          interceptor._handle_grpc_failure,
-                          MockRpcErrorResponse())
-
-    def test_handle_grpc_failure_not_google_ads_failure(self):
-        """Raises as-is non-retryable non-GoogleAdsFailure exceptions."""
-        class MockRpcErrorResponse(grpc.RpcError):
-            def code(self):
-                return grpc.StatusCode.INVALID_ARGUMENT
-
-            def trailing_metadata(self):
-                return (('bad-failure-key', 'arbitrary-value'),)
-
-            def exception(self):
-                return self
-
-        interceptor = self._create_test_interceptor()
-
-        self.assertRaises(MockRpcErrorResponse,
-                          interceptor._handle_grpc_failure,
-                          MockRpcErrorResponse())
-
-    def test_intercept_unary_unary_response_is_exception(self):
-        """If response.exception() is not None exception is handled."""
-        mock_exception = grpc.RpcError()
-
-        class MockResponse():
-            def exception(self):
-                return mock_exception
-
-        mock_request = mock.Mock()
-        mock_client_call_details = mock.Mock()
-        mock_response = MockResponse()
-
-        def mock_continuation(client_call_details, request):
-            del client_call_details
-            del request
-            return mock_response
-
-        interceptor = self._create_test_interceptor()
-
-        with mock.patch.object(interceptor, '_handle_grpc_failure'):
-            interceptor.intercept_unary_unary(
-                mock_continuation, mock_client_call_details, mock_request)
-
-            interceptor._handle_grpc_failure.assert_called_once_with(
-                mock_response)
-
-    def test_intercept_unary_unary_response_is_successful(self):
-        """If response.exception() is None response is returned."""
-        class MockResponse():
-            def exception(self):
-                return None
-
-        mock_request = mock.Mock()
-        mock_client_call_details = mock.Mock()
-        mock_response = MockResponse()
-
-        def mock_continuation(client_call_details, request):
-            del client_call_details
-            del request
-            return mock_response
-
-        interceptor = self._create_test_interceptor()
-
-        result = interceptor.intercept_unary_unary(
-            mock_continuation, mock_client_call_details, mock_request)
-
-        self.assertEqual(result, mock_response)
